requirements.txt
# ============================================================================
# TENBOT - Ultra Discord Bot for Business Communities
# ============================================================================
# Python package requirements
# Install with: pip install -r requirements.txt

# Core Discord
discord.py>=2.3.2
aiohttp>=3.9.0
discord-ext-voice-recv>=0.4.0

# Database
aiosqlite>=0.19.0

# Image Processing (for fingerprinting spam images)
Pillow>=10.0.0
imagehash>=4.3.1

# Utilities
python-dotenv>=1.0.0

# Audio transcription (local)
faster-whisper>=1.0.0

# Future AI Integration (commented out for now)
# anthropic>=0.18.0
# openai>=1.12.0
